Model, schemas, and doc fixes for Item.
This makes use of the `pymdownx.details` extension to allow for
collapsed/folded code example blocks. (Other commits/PRs to come will
also use this, but this is the first set.) I also replaced some uses of
`MyModel` and `YourModel` (and associated schemas) with `Item` (and
etc.) for simplicity, although there will also be some `MyModel` usage
in the following commit.